+++ middlewares/uploadImages.js
@@ -3,7 +3,6 @@ const Jimp = require('jimp');
 const path = require('path');
 const fs = require('fs');
 
-
 const multerStorage = multer.diskStorage({
     destination: function(req, file, cb) {
         cb(null, path.join(__dirname, '../public/images'));
@@ -18,9 +17,7 @@ const multerFilter = (req, file, cb) => {
     if (file.mimetype.startsWith('image')) {
         cb(null, true);
     } else {
-        cb({
-            message: 'Unsupported file format'
-        });
+        cb({ message: 'Unsupported file format' });
     }
 };
 
@@ -30,23 +27,31 @@ const uploadPhoto = multer({
     limits: { fieldNameSize: 2000000 }
 });
 
-
 const productImgResize = async (req, res, next) => {
     if (!req.files || !Array.isArray(req.files)) return next();
+
     await Promise.all(
         req.files.map(async (file) => {
             try {
                 const img = await Jimp.read(file.path);
-                await img.resize(300, 300).quality(90).write(`public/images/products/${file.filename}`);
-                fs.unlinkSync(`public/images/products/${file.filename}`);
+                await img.resize(300, 300).quality(90).writeAsync(`public/images/products/${file.filename}`);
+                console.log("Imagen redimensionada y escrita correctamente");
+
+                // Eliminar el archivo
+                if (fs.existsSync(`public/images/products/${file.filename}`)) {
+                    fs.unlinkSync(`public/images/products/${file.filename}`);
+                    console.log("Imagen eliminada correctamente");
+                } else {
+                    console.warn(`El archivo no existe: public/images/products/${file.filename}`);
+                }
             } catch (error) {
                 console.error("Error al procesar la imagen:", error);
             }
         })
     );
+
     next();
 };
-
 
 
 const blogImgResize = async (req, res, next) => {
